+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 ; logging options
-log_cli = False
+log_cli = True
 addopts = --verbose --tb=short -n auto -p no:monitor
 log_level = WARNING
 console_output_style = count
